Rewrite NewPredict
+ form of date, time, and select model id
